split programs
